add the ChangingSwimmer class
Add the ChangingSwimmer class, which extends the PaintedCreature class.
It is essentially a clone of the LineSwimmer with get and set methods
for deltaX and deltaY; I decided to make it a new class because it is
easier to make sure it has been added to the students' local copies of
the aquarium project, compared to having them overwrite the previous
file.